feat(backend): :sparkles: Add JWT token generation and verification
Added JWT token generation, verification and revocation to the backend.
This will be used to authenticate users and protect routes.

The token will be signed with a secret key and a random string that changes every restart of the server.
This will affect all the tokens that are valid but it will protect the server from attacks.

closes #43