feat(frontend): :sparkles: Add survey management
Added the ability to manage surveys, including creating, editing the votes (add, remove).
TODO: Add the ability to delete a survey and change the status to closed.

Added assets for the voting page for the votes from 1 to 5.

Changed some types according API documentation
Modified component `ListaSondaggi` for better styling
Modified utilities files for better handling of data
Modified `App.vue` with the confirmService for the delete/close buttons